Refactor common SQL methods
